refactor(roast-timer): RoastTimer.tsx を責務別に分割
RoastTimer.tsx (1,421行) を以下のコンポーネントに分割:

- constants.ts: ROAST_LEVELS, WEIGHTS, DEFAULT_DURATION_BY_WEIGHT
- utils.ts: convertToHalfWidth, removeNonNumeric
- TimerDisplay.tsx: 円形プログレスバーと時間表示
- TimerControls.tsx: 操作ボタン（一時停止/再開/スキップ/リセット）
- TimerHeader.tsx: ヘッダーボタン（戻る/設定/記録一覧）
- SetupPanel.tsx: 設定フォーム（idle状態）
- index.ts: エクスポート

分割後:
- RoastTimer.tsx: 402行（71%削減）
- 合計: 1,479行（8ファイル）

Closes #34
